requirements.txt
prompt_toolkit
pydantic
pyyaml
sqlalchemy
rich
